// src/types/mock-data.type.ts
export type MockDataType = {
  name: string[];
  description: string[];
  genre: string[];
  director: string[];
  starring: string[];
  posterImage: string[];
  backgroundImage: string[];
  backgroundColor: string[];
  previewVideoLink: string[];
  videoLink: string[];
  username: string[];
  email: string[];
  avatarPath: string[];
}